Use Mojo::UserAgent::CookieJar behavior from Mojolicious >= 6.20
In the past several years, Mojo::UserAgent::CookieJar has changed the
method used to disable it several times. With Mojolicious >= 5.45, the
previous $ua->cookie_jar(0) caused an error like:

   Can't locate object method "prepare" via package "0"

At that point, the way to disable it was changed to:
   $ua->cookie_jar->extracting(0).

With Mojolicious 6.0, the behavior changed yet again - it became:
   $ua->cookie_jar->collecting(0);

Finally, it has stabilized since 6.20 and until current (8.12) to:
   $ua->cookie_jar->ignore(sub { 1 });

Rather than add something like conditionals for $Mojolicious::VERSION
and support every deprecated method, this commit just changes all the
dependencies for Mojolicious to >= 6.20 and uses that behavior.

Due to the potentially breaking changes for those on older versions of
Mojolicious, the version number of the module has been bumped to 1.0.2.

This also includes a new test, subscribe.t, which attempts to make a
subscription to test.invalid and catches any errors in an eval block. An
alarm() and associated handler are used to leave the ->subscribe loop. I
wasn't sure if it'd be suitable for Windows due to the emulated SIGALRM,
but it passes there as well.

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>